feat: update docker-compose with latest librechat images
Updates docker-compose.yml to use the latest images for librechat, meilisearch, and vectordb.

BREAKING CHANGE: This change removes the chat-meilisearch_data docker volume. If you are upgrading from a previous version of librechat, you will need to remove the chat-meilisearch_data volume before restarting docker-compose.

Additionally, this change introduces a new environment variable, MEILI_HOST, which must be set to the hostname or IP address of the meilisearch container.